requirements_eval.txt
# --- Core Configuration ---
python-dotenv>=1.0.0
azure-identity>=1.15.0

# --- Azure AI Project & Agent Services ---
azure-ai-projects>=1.0.0b1
azure-ai-agents>=0.1.0b1

# --- Evaluation Framework SDK ---
azure-ai-evaluation>=1.0.0

# --- Core LLM Client ---
openai>=1.12.0

# --- Data Processing & Reporting ---
pandas>=2.1.4
jinja2>=3.1.2
pyyaml>=6.0.1
